# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Verification code sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpSentResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Phone number already registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Register with name/phone/password and request a verification code",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/resend-registration-code": {
      "post": {
        "operationId": "AuthController_resendRegistrationCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendRegistrationCodeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Verification code resent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpSentResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "No pending registration found for this phone number",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Resend the registration verification code",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/verify-registration": {
      "post": {
        "operationId": "AuthController_verifyRegistration",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyRegistrationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account activated; returns access + refresh tokens",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Verify the code and activate the account",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Returns access + refresh tokens",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Login with phone number and password",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Code sent (or silently ignored if phone not found)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpSentResponseDto"
                }
              }
            }
          }
        },
        "summary": "Request a password-reset verification code",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Password updated; returns new tokens",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Reset password using the verification code",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Returns new access + refresh tokens",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid refresh token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Refresh access token using a refresh token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Logged out successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revoke refresh token and logout",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/v1/admin/users": {
      "get": {
        "operationId": "AdminController_listUsers",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (default 1)",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (1-100, default 20)",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Pagination offset (0-10000, default 0)",
            "schema": {
              "minimum": 0,
              "maximum": 10000,
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by account status",
            "schema": {
              "example": "BLOCKED",
              "type": "string",
              "enum": [
                "PENDING_VERIFICATION",
                "ACTIVE",
                "BLOCKED"
              ]
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Search by name or phone (1–100 chars)",
            "schema": {
              "minLength": 1,
              "maxLength": 100,
              "example": "Ahmed",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated user list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUsersListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List users with optional filters (admin only)",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/users/{id}": {
      "get": {
        "operationId": "AdminController_getUserDetails",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user details (admin only)",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/users/{id}/status": {
      "patch": {
        "operationId": "AdminController_updateUserStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User status updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a user's status, e.g. to block them (admin only)",
        "tags": [
          "Admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "BW Café Ordering API",
    "description": "Single-restaurant delivery/pickup ordering API — guest and registered checkout, admin order management.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 150,
            "description": "Full name (2–150 chars)",
            "example": "Ahmed Ali"
          },
          "phone": {
            "type": "string",
            "minLength": 7,
            "maxLength": 32,
            "description": "Phone number (7–32 chars)",
            "example": "+22200000000"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 64,
            "pattern": "/^(?=.*[A-Za-z])(?=.*\\d).+$/",
            "description": "Password — must contain letters and numbers (8–64 chars)",
            "example": "Secret123"
          }
        },
        "required": [
          "name",
          "phone",
          "password"
        ]
      },
      "OtpSentDataDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Verification code sent"
          },
          "code": {
            "type": "string",
            "example": "000000",
            "description": "No SMS provider is wired up yet — the code is returned here instead of being sent."
          }
        },
        "required": [
          "message"
        ]
      },
      "OtpSentResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/OtpSentDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
      "ErrorDetailDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 409
          },
          "message": {
            "type": "string",
            "example": "Conflict"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-03-28T12:00:00.000Z"
          },
          "path": {
            "type": "string",
            "example": "/api/auth/register/request-otp"
          }
        },
        "required": [
          "code",
          "message",
          "timestamp",
          "path"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "statusCode": {
            "type": "number",
            "example": 409
          },
          "data": {
            "type": "object",
            "nullable": true,
            "additionalProperties": false,
            "example": null
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetailDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data",
          "error"
        ]
      },
      "ResendRegistrationCodeDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "minLength": 7,
            "maxLength": 32,
            "description": "Phone number",
            "example": "+22200000000"
          }
        },
        "required": [
          "phone"
        ]
      },
      "VerifyRegistrationDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "minLength": 7,
            "maxLength": 32,
            "description": "Phone number",
            "example": "+22200000000"
          },
          "code": {
            "type": "string",
            "minLength": 4,
            "maxLength": 8,
            "description": "One-time code sent to the phone (4–8 digits)",
            "example": "000000"
          }
        },
        "required": [
          "phone",
          "code"
        ]
      },
      "AuthUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Ahmed Ali"
          },
          "phone": {
            "type": "string",
            "example": "+22200000000"
          },
          "role": {
            "type": "string",
            "example": "USER",
            "enum": [
              "USER",
              "ADMIN"
            ]
          },
          "status": {
            "type": "string",
            "example": "ACTIVE",
            "enum": [
              "PENDING_VERIFICATION",
              "ACTIVE",
              "BLOCKED"
            ]
          }
        },
        "required": [
          "id",
          "name",
          "phone",
          "role",
          "status"
        ]
      },
      "TokenDataDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AuthUserDto"
          },
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "refreshToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "accessToken",
          "refreshToken"
        ]
      },
      "TokenResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/TokenDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "minLength": 7,
            "maxLength": 32,
            "description": "Phone number",
            "example": "+22200000000"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 64,
            "description": "User password (8–64 chars)",
            "example": "Secret123"
          }
        },
        "required": [
          "phone",
          "password"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "minLength": 7,
            "maxLength": 32,
            "description": "Registered phone number",
            "example": "+22200000000"
          }
        },
        "required": [
          "phone"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "minLength": 7,
            "maxLength": 32,
            "description": "Phone number associated with the account",
            "example": "+22200000000"
          },
          "code": {
            "type": "string",
            "minLength": 4,
            "maxLength": 8,
            "description": "One-time code received (4–8 digits)",
            "example": "000000"
          },
          "newPassword": {
            "type": "string",
            "minLength": 8,
            "maxLength": 64,
            "pattern": "/^(?=.*[A-Za-z])(?=.*\\d).+$/",
            "description": "New password — must contain letters and numbers (8–64 chars)",
            "example": "NewSecret123"
          },
          "confirmPassword": {
            "type": "string",
            "minLength": 8,
            "maxLength": 64,
            "description": "Must match newPassword",
            "example": "NewSecret123"
          }
        },
        "required": [
          "phone",
          "code",
          "newPassword",
          "confirmPassword"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "JWT refresh token",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "LogoutDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "The refresh token to revoke",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "LogoutDataDto": {
        "type": "object",
        "properties": {}
      },
      "LogoutResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/LogoutDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
...
      "AdminUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Ahmed Mohamed"
          },
          "phone": {
            "type": "string",
            "example": "+22200000000"
          },
          "role": {
            "type": "string",
            "example": "USER",
            "enum": [
              "USER",
              "ADMIN"
            ]
          },
          "status": {
            "type": "string",
            "example": "ACTIVE",
            "enum": [
              "PENDING_VERIFICATION",
              "ACTIVE",
              "BLOCKED"
            ]
          }
        },
        "required": [
          "id",
          "name",
          "phone",
          "role",
          "status"
        ]
      },
      "AdminUsersListDataDto": {
        "type": "object",
        "properties": {
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminUserDto"
            }
          }
        },
        "required": [
          "users"
        ]
      },
      "AdminUsersListResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/AdminUsersListDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
      "AdminUserDataDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AdminUserDto"
          }
        },
        "required": [
          "user"
        ]
      },
      "AdminUserResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/AdminUserDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
      "UpdateUserStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "BLOCKED"
            ],
            "description": "New account status for the user",
            "example": "BLOCKED"
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}